fix: MovieStatsResponse averageRating 필드 타입 수정 - Long -> Double로 변경하여 소수점 평점 지원 - MovieService와의 타입 호환성 확보

+++ backend/src/main/java/com/moviebuddies/dto/response/MovieStatsResponse.java
@@ -26,8 +26,8 @@ public class MovieStatsResponse {
     private long nowPlayingMovies;
 
     /**
-     * 전체 영화으 평균 평점 (TMDB 기준)
+     * 전체 영화의 평균 평점 (TMDB 기준)
      */
-    private long averageRating;
+    private Double averageRating;
 
 }
